Adding final code to create the GA4 tag.
Also made a change to consolidate all of the config gathering from the form to
reduce the number of times the DOM is touched.

Disabled the form during deployment to avoid people thinking they can change
things during the process.

Added printing success and failure messages to the web page.